Added model05,06, with reprojected targets.
Since previous models were having trouble with warped textures on
surfaces, I put in some generated reprojected images based on the
generated (warped) image, that the network should try to create.

Also moved the position injection from the start of the encoder, to
the start of the decoder. Also upscale it and add it at all skip
connection layers.

Added Spectral Norm to 05,06, but haven't tested it.