Model08 to try to improve training.
Model 08 refactored how resolution works, making the model almost
entirely independent of resolution. Changed dataloader to move the
parameters into the loader rather than the transform constructors.
Moved the final resample outside the dataloader, do it at the start
of each batch. That way each batch can be dynamically resized.
This is slower, because the dataloader is more parallelized, but it
will hopefully make the model less susceptible to issues when switching
between resolutions.