Removed batchnorm, fixed w0 in sin.
Checked out siren-pytorch, and noted that I had these things wrong.
No batchnorm (it was causing deeper networks to converge super slow, if
at all). Changed the w0 factor to be in the sin activation, instead of
just scaling the weight initially.

Also added print-statistics, to see basic mean/stdev of weights for all
layers, to see if/when they change.