Added position randomization, and renderer.
Position randomization transform is applied during training. For an
input position+direction, randomly select a offset along the view
direction so it's still within the +/- 1 bounding box of the camera.
Tried implementing it as a dataset transform, but the way I have the
datasets split up, each sample is a single pixel, and doing the math is
super inefficient. Vectorized it and moved it to do entire batches.

Fixed the theta/phi of inputs; I was applying a sin to the -1..1 values,
which are already sin values, effectively shrinking the range of those
inputs.

Also added a render_demo, which is a little OpenGL/GLUT app that takes
a model, and can render views. Not perspective warped yet, and very
inefficient (copy pytorch tensor from GPU to CPU, then copy back to GPU
OpenGL texture for rendering with a screen space oriented quad).
Still, it works.