Changed v4 image offsets.
Was using half pixel angle from the center pixel of the image. But it
could be much larger. The edge pixels are not at 90 deg, but at
asin(1-1/w), which for a 512x512 image is 86.4 deg. So let the views
vary by up to what will cause the edge pixels to be 90.

Also changed default scheduler gamma to not have any effect.