Added importance sampling to dataloader.
Changed how dataloader samples are indexed. Instead of calculating them,
explicitly make a tensor with a row for each index (sub-image, x, y).
But this also allows me to repeat indices. Use a laplassian on the image
to find edges that would benefit from more sampling, do some arbitrary
math to get a count for each sample, and repeat_interleave the important
indices.