More refactoring of dataloaders.
The outer dataloader, for the FileListLoader dataset, now copies the
images/tensors to the GPU. The inner dataloader is now a stand alone
class, rather than a dataset working with the Pytorch dataloader.

Because all the relevant tensors are on the GPU, I vectorized many of
the calculations (index calc), implemented a shuffle with the importance
sample augmented indices, and moved the input vector transform into the
dataloader __getitem__.

Added a basic profiling script, which I used to help figure out where
time was being spent in training.