Optimized some cats, made model 03.
From profiling, a lot of time was spent in cat. To improve this, I
changed a few thing. In the sample dataloader init, instead of catting
the tensors as the list is parsed, put them all into lists and cat them
once at the end of the loop.

The pos_encoding step used loops and nested cats to turn inputs into the
expanded position encoding. Refactored it to do it with tensors. This
re-arranged the order of the inputs, so I incremented the model number.