Fixed rotation calculation in dataloader.
Turns out naively adding rotation to theta/phi is not correct.
It's now done with a rotation of the view vectors. Noticed that if I
want resulting vector to be in the -Z half-space, the limit on phi
rotation is very limited, because vertical rotation warps the corners
outside the space. Changed how random value is generated to keep it in
range.

Changed the filenames a bit, rotation now scientific notation.

Also added sample skew to the output of the loss.