Added auto resume functionality.
Changed save model to save to a single resume/checkpoint file. If one
startup a file is found, it will use that. However, you can still
specify another file to load, and it will start from that instead.
Now the main run command is in a loop with a try/except. If an exception
is caught, it'll retry the command, which will resume from the last
checkpoint.

Tweaked the random seed, to accomodate the resume. Now the manual seed
is set every epoch, based on a seed passed in (if available), or its
seeded randomly if not.

Also added DistributedSampler for test set.